+++ polling_stations/apps/data_collection/tests/test_address_list.py
@@ -1,4 +1,4 @@
-from data_collection.management.commands import AddressList
+from data_collection.custom_lists import AddressList
 from django.test import TestCase
